[#2118]: WIP: `Default` config
