+++ config/base/Cargo.toml
@@ -12,3 +12,4 @@ serde_json = "1.0"
 thiserror = "1.0.30"
 crossbeam = "0.8.1"
 derive_more = "0.99.17"
+eyre = "0.6.8"
